Add files via upload
Currently, I'm able to read the url for that specific fanart and able to extract the title of it. What I'm working on now is to figure out a way to move that file to a folder, which I need to find a way to create a new folder if it doesn't exist.